moved tests for mrdata submodule into single test file and added some tests covering RuntimeErrors of setter functions.
